Separate Winston injections
This separates Winston's injections for the base game and Golden Mask, and the nose texture fix in OG gym and Vegas.
