+++ src/Types/TR2/Textures/TR2VegasTextureBuilder.cs
@@ -4,28 +4,24 @@ using TRXInjectionTool.Control;
 
 namespace TRXInjectionTool.Types.TR2.Textures;
 
-public class TR2GymTextureBuilder : TextureBuilder
+public class TR2VegasTextureBuilder : TextureBuilder
 {
-    public override string ID => "gym_textures";
+    public override string ID => "vegas_textures";
 
     public override List<InjectionData> Build()
     {
         InjectionData data = CreateBaseData();
-
-        TR2Level gym = _control2.Read($"Resources/{TR2LevelNames.ASSAULT}");
-        FixLaraTransparency(gym, data);
-
         return new() { data };
     }
 
     private InjectionData CreateBaseData()
     {
-        TR2Level level = CreateWinstonLevel(TR2LevelNames.ASSAULT);
+        TR2Level level = CreateWinstonLevel(TR2LevelNames.VEGAS);
         // Current injection limitation, do not replace SFX
         level.SoundEffects.Clear();
 
         InjectionData data = InjectionData.Create(level, InjectionType.TextureFix, ID);
-        CreateDefaultTests(data, TR2LevelNames.ASSAULT);
+        CreateDefaultTests(data, TR2LevelNames.VEGAS);
         return data;
     }
 }
